Struct oxischeme::eval::Meaning

pub struct Meaning {
    // some fields omitted
}

The Meaning type is our intermediate language produced by syntactic analysis. It is a triple containing a MeaningData variant, its corresponding MeaningEvaluatorFn, and the source location this Meaning originates from.