Struct oxischeme::heap::Arena

pub struct Arena<T> {
    // some fields omitted
}

An arena from which to allocate T objects from.

Methods

impl<T: Default> Arena<T>

fn new(capacity: uint) -> Box<Arena<T>>

Create a new Arena with the capacity to allocate the given number of T instances.

impl<T> Arena<T>

fn capacity(&self) -> uint

Get this heap's capacity for simultaneously allocated cons cells.

fn is_full(&self) -> bool

Return true if this arena is at full capacity, and false otherwise.

fn allocate(&mut self) -> ArenaPtr<T>

Allocate a new T instance and return a pointer to it.

Panics

Panics if the number of T instances allocated is already equal to this Arena's capacity and no more T instances can be allocated.

fn sweep(&mut self, live: HashSet<uint>)

Sweep the arena and add any reclaimed objects back to the free list.