Struct oxischeme::heap::ArenaSet

pub struct ArenaSet<T> {
    // some fields omitted
}

A set of Arenas. Manages allocating and deallocating additional Arenas from the OS, depending on the number of objects requested and kept alive by the mutator.

Methods

impl<T: Default> ArenaSet<T>

fn new(capacity: usize) -> ArenaSet<T>

Create a new ArenaSet.

fn sweep(&mut self)

Sweep all of the arenas in this set.

fn allocate(&mut self) -> ArenaPtr<T>

Allocate a T object from one of the arenas in this set and return a pointer to it.