Struct oxischeme::heap::Heap

pub struct Heap {
    // some fields omitted
}

The scheme heap and GC runtime, containing all allocated cons cells, environments, procedures, and strings (including strings for symbols).

Methods

impl Heap

Heap Constructors

fn new() -> Heap

Create a new Heap with the default capacity.

fn with_arenas(cons_cells: Box<Arena<Cons>>, strings: Box<Arena<String>>, envs: Box<Arena<Environment>>, procs: Box<Arena<Procedure>>) -> Heap

Create a new Heap using the given arenas for allocating cons cells and strings within.

impl Heap

Heap Allocation Methods

fn allocate_cons(&mut self) -> RootedConsPtr

Allocate a new cons cell and return a pointer to it.

Panics

Panics if the Arena for cons cells has already reached capacity.

fn allocate_string(&mut self) -> RootedStringPtr

Allocate a new string and return a pointer to it.

Panics

Panics if the Arena for strings has already reached capacity.

fn allocate_environment(&mut self) -> RootedEnvironmentPtr

Allocate a new Environment and return a pointer to it.

Panics

Panics if the Arena for environments has already reached capacity.

fn allocate_procedure(&mut self) -> RootedProcedurePtr

Allocate a new Procedure and return a pointer to it.

Panics

Panics if the Arena for environments has already reached capacity.

impl Heap

Heap Methods for Garbage Collection

fn collect_garbage(&mut self)

Perform a garbage collection on the heap.

fn add_root(&mut self, root: GcThing)

Explicitly add the given GC thing as a root.

fn drop_root(&mut self, root: GcThing)

Unroot a GC thing that was explicitly rooted with add_root.

fn increase_gc_pressure(&mut self)

Apply pressure to the GC, and if enough pressure has built up, then perform a garbage collection.

impl Heap

Heap Methods and Accessors.

fn global_env(&mut self) -> RootedEnvironmentPtr

Get the global environment.

fn get_or_create_symbol(&mut self, str: String) -> RootedValue

Ensure that there is an interned symbol extant for the given String and return it.

impl Heap

Getters for well known symbols.

fn quote_symbol(&mut self) -> RootedValue

fn if_symbol(&mut self) -> RootedValue

fn begin_symbol(&mut self) -> RootedValue

fn define_symbol(&mut self) -> RootedValue

fn set_bang_symbol(&mut self) -> RootedValue

fn unspecified_symbol(&mut self) -> RootedValue

fn lambda_symbol(&mut self) -> RootedValue