Struct oxischeme::read::Read

pub struct Read<R: Reader> {
    // some fields omitted
}

Read iteratively parses values from the input Reader.

Methods

impl<'a, R: Reader> Read<R>

fn new(reader: R, heap: *mut Heap, file_name: String) -> Read<R>

Create a new Read instance from the given Reader input source.

fn get_result(&'a self) -> &'a Result<(), String>

Get the results of parsing thus far. If there was an error parsing, a diagnostic message will be the value of the error.