Struct oxischeme::value::ConsIterator

pub struct ConsIterator {
    // some fields omitted
}

An iterator which yields Ok for each value in a cons-list and finishes with None when the end of the list is reached (the scheme empty list value) or Err when iterating over an improper list.

For example: the list (1 2 3) would yield Ok(1), Ok(2), Ok(3), None, while the improper list (1 2 . 3) would yield Ok(1), Ok(2), Err.