Struct oxischeme::value::Procedure

pub struct Procedure {
    pub arity: u32,
    pub body: Option<Box<Meaning>>,
    pub act: Option<ActivationPtr>,
}

User defined procedures are represented by their body and a pointer to the activation that they were defined within.

Fields

arity

body

act