Struct oxischeme::environment::Environment

pub struct Environment {
    // some fields omitted
}

The Environment associates symbols with values.

Methods

impl Environment

fn extend(heap: &mut Heap, parent: &RootedEnvironmentPtr, names: &RootedValue, values: &RootedValue) -> Result<RootedEnvironmentPtr, String>

Extend the given environment with the names and associated values supplied, resulting in a new environment.

fn set_parent(&mut self, parent: &RootedEnvironmentPtr)

Set the parent of this environment. When looking up bindings, if this environment doesn't have the target binding, and this environment has a parent environment, we will recurse to the parent and do a lookup in that environment, and so on until either there are no more environments or we find the binding.

fn define(&mut self, sym: String, val: &RootedValue)

Define a new variable bound to the given value.

fn define_unrooted(&mut self, sym: String, val: Value)

Define a new variable bound to the given (unrooted) value.

fn update(&mut self, sym: String, val: &RootedValue) -> Result<(), String>

Update an existing binding to be associated with the new value.

fn lookup(&self, heap: &mut Heap, sym: &String) -> SchemeResult

Lookup the value associated with the given symbol.